requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
azure-ai-formrecognizer==3.3.2
azure-core==1.29.5
azure-cosmos==4.5.1
openai==1.3.7
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
httpx>=0.28.0